new css designs

+++ workforce-backend/src/modules/scheduling/shiftTemplate/shiftTemplate.service.ts
@@ -1,12 +1,13 @@
 import Schedule from "../../../models/schedule.model";
+import ShiftSession from "../../../models/ShiftSession";
 import ShiftTemplate from "../../../models/shiftTemplate.model";
 
 export const createTemplate = async (data: any) => {
   return await ShiftTemplate.create(data);
 };
 
-export const getTemplates = async () => {
-  return await ShiftTemplate.find();
+export const getTemplates = async (includeInactive = false) => {
+  return await ShiftTemplate.find(includeInactive ? {} : { isActive: { $ne: false } });
 };
 
 export const getTemplateById = async (id: string) => {
@@ -19,9 +20,14 @@ export const updateTemplate = async (id: string, data: any) => {
 
 export const deleteTemplate = async (id: string) => {
   const assignedSchedule = await Schedule.findOne({ shiftTemplateId: id });
+  const linkedSession = await ShiftSession.findOne({ shiftTemplateId: id });
 
-  if (assignedSchedule) {
-    throw new Error("Cannot delete a template assigned to schedules");
+  if (assignedSchedule || linkedSession) {
+    return await ShiftTemplate.findByIdAndUpdate(
+      id,
+      { isActive: false },
+      { new: true }
+    );
   }
 
   return await ShiftTemplate.findByIdAndDelete(id);
